+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ** 1.0.0 **
 
 * Initial version
+
+** 1.0.1 **
+
+* Implemented Capture on Complete feature
+* Implemented Void on Cancelled feature
+* Updated translations